+++ libs/langgraph-core/src/tests/pregel.test.ts
@@ -9188,7 +9188,7 @@ graph TD;
       graph.stream({ messages: [] }, { ...config, streamMode: "messages" })
     );
 
-    expect(streamedEvents).toEqual([
+    expect(streamedEvents).toMatchObject([
       [
         new _AnyIdToolMessage({
           tool_call_id: "test",
@@ -9253,7 +9253,7 @@ graph TD;
           langgraph_checkpoint_ns: expect.stringMatching(/^p_two:.*\|c_two:.*/),
           __pregel_task_id: expect.any(String),
           checkpoint_ns: expect.stringMatching(/^p_two:/),
-          ls_integration: "langgraph",
+          ls_integration: "langchain_chat_model",
           ls_model_type: "chat",
           ls_provider: "FakeChatModel",
           ls_stop: undefined,
@@ -9274,7 +9274,7 @@ graph TD;
           langgraph_checkpoint_ns: expect.stringMatching(/^p_two:.*\|c_two:.*/),
           __pregel_task_id: expect.any(String),
           checkpoint_ns: expect.stringMatching(/^p_two:/),
-          ls_integration: "langgraph",
+          ls_integration: "langchain_chat_model",
           ls_model_type: "chat",
           ls_provider: "FakeChatModel",
           ls_stop: undefined,
@@ -9295,7 +9295,7 @@ graph TD;
           langgraph_checkpoint_ns: expect.stringMatching(/^p_three/),
           __pregel_task_id: expect.any(String),
           checkpoint_ns: expect.stringMatching(/^p_three/),
-          ls_integration: "langgraph",
+          ls_integration: "langchain_chat_model",
           ls_model_type: "chat",
           ls_provider: "FakeChatModel",
           ls_stop: undefined,
@@ -9410,7 +9410,7 @@ graph TD;
               expect.stringMatching(/^p_two:.*\|c_two:.*/),
             __pregel_task_id: expect.any(String),
             checkpoint_ns: expect.stringMatching(/^p_two:/),
-            ls_integration: "langgraph",
+            ls_integration: "langchain_chat_model",
             ls_model_type: "chat",
             ls_provider: "FakeChatModel",
             ls_stop: undefined,
@@ -9432,7 +9432,7 @@ graph TD;
               expect.stringMatching(/^p_two:.*\|c_two:.*/),
             __pregel_task_id: expect.any(String),
             checkpoint_ns: expect.stringMatching(/^p_two:/),
-            ls_integration: "langgraph",
+            ls_integration: "langchain_chat_model",
             ls_model_type: "chat",
             ls_provider: "FakeChatModel",
             ls_stop: undefined,
@@ -9452,7 +9452,7 @@ graph TD;
             langgraph_checkpoint_ns: expect.stringMatching(/^p_three/),
             __pregel_task_id: expect.any(String),
             checkpoint_ns: expect.stringMatching(/^p_three/),
-            ls_integration: "langgraph",
+            ls_integration: "langchain_chat_model",
             ls_model_type: "chat",
             ls_provider: "FakeChatModel",
             ls_stop: undefined,